# use exponential search to speed up lexico partition